api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
uvicorn
fastapi
python-multipart
pypdf
cryptography>=3.1
openai-whisper
PyMuPDF
whisper
yt-dlp